requirements.txt
customtkinter
moviepy==1.0.3
openai-whisper
requests
packaging
colorlog
python-dotenv
google-generativeai
soundfile
torch
numpy
pillow
imageio_ffmpeg
PySide6>=6.5.0